+++ packages/browser/src/image/convert.ts
@@ -5,8 +5,9 @@ import {
   getBlobFromCanvasAsync,
 } from "../canvas/convert";
 import { bytesToBase64 } from "@a2-ts-utils/common/byte";
+import { getImageElementAsync } from "./image";
 
-//
+// element > blob =====
 export const getBlobFromElementAsync = (
   element: HTMLVideoElement | HTMLImageElement | HTMLCanvasElement,
   type: ImageMimeType = ImageMimeType.JPEG,
@@ -18,6 +19,7 @@ export const getBlobFromElementAsync = (
 
 export const intoElementToBlobAsync = getBlobFromElementAsync;
 
+// element (> blob) > bytes =====
 export const getBytesFromElementAsync = async (
   element: HTMLVideoElement | HTMLImageElement | HTMLCanvasElement,
   type: ImageMimeType = ImageMimeType.JPEG,
@@ -29,6 +31,7 @@ export const getBytesFromElementAsync = async (
 
 export const intoElementToBytesAsync = getBytesFromElementAsync;
 
+// bytes > base64 =====
 export const getBase64SrcFromBytes = (
   bytes: Uint8Array,
   type: ImageMimeType = ImageMimeType.JPEG,
@@ -38,3 +41,50 @@ export const getBase64SrcFromBytes = (
 };
 
 export const intoBytesToBase64Src = getBase64SrcFromBytes;
+
+// resize blob image
+// mainly work with set canvas
+export const resizeBlob = async (
+  blob: Blob,
+  type: ImageMimeType = ImageMimeType.JPEG,
+  newSize: Size,
+): Promise<Blob> => {
+  const url = URL.createObjectURL(blob);
+  try {
+    const imageElement = await getImageElementAsync(url);
+    return await getBlobFromElementAsync(imageElement, type, newSize);
+  } finally {
+    URL.revokeObjectURL(url);
+  }
+};
+
+// resize bytes image
+// bytes > blob > resizeBlob > blob
+export const resizeBytes = async (
+  bytes: Uint8Array,
+  type: ImageMimeType = ImageMimeType.JPEG,
+  newSize: Size,
+): Promise<Blob> => {
+  const blob = getBlobFromBytes(bytes);
+  return await resizeBlob(blob, type, newSize);
+};
+
+// blob > element
+export const getImageElementFromBlob = async (
+  blob: Blob,
+): Promise<HTMLImageElement> => {
+  const url = URL.createObjectURL(blob);
+  try {
+    return await getImageElementAsync(url);
+  } finally {
+    URL.revokeObjectURL(url);
+  }
+};
+
+export const getBlobFromBytes = (bytes: Uint8Array) => {
+  const arrayBuffer = bytes.buffer.slice(
+    bytes.byteOffset,
+    bytes.byteOffset + bytes.byteLength,
+  ) as ArrayBuffer;
+  return new Blob([arrayBuffer]);
+};