+++ packages/common/src/index.ts
@@ -3,5 +3,6 @@ export * as image from "./image";
 export * as type from "./type";
 export * as byte from "./byte";
 export * as data from "./data";
+export * as time from "./time";
 
 export const Name = "@a2-ts-utils/common";