+++ packages/browser/history.md
@@ -2,6 +2,8 @@
 
 ## 26-05
 
+- 05-20(wed)
+  - add `resizeBlob`, `resizeBytes`
 - 05-06(wed)
   - image, canvasのメソッド名を修正
     - getXXFromYYY, intoYYYToXXXという名前で統一することにした